New print family from musl

fprintf/vprintf from musl. Fprintf dies at vfprintf:672
testwchar dies at pwrite syscall.

removed syscall wrappers: chdir, open, close, pipe.
All tested and working.

Updated ap/internal fcall.h according to native fcall.h
(fids, etc..)

Updated ap/plan9/profile.c (unused)

Fixed size_t

+++ lib/ap/internal/fcall.h
@@ -30,7 +30,7 @@ enum
 struct	Fcall
 {
 	char	type;
-	int16_t	fid;
+	uint32_t	fid;
 	uint16_t	tag;
 	union
 	{
@@ -56,15 +56,15 @@ struct	Fcall
 		};
 		struct
 		{
-			int32_t	perm;		/* T-Create */ 
-			int16_t	newfid;		/* T-Clone, T-Clwalk */
+			uint32_t	perm;		/* T-Create */ 
+			uint32_t	newfid;		/* T-Clone, T-Clwalk */
 			char	name[NAMELEN];	/* T-Walk, T-Clwalk, T-Create */
 			char	mode;		/* T-Create, T-Open */
 		};
 		struct
 		{
-			int32_t	offset;		/* T-Read, T-Write */
-			int32_t	count;		/* T-Read, T-Write, R-Read */
+			int64_t	offset;		/* T-Read, T-Write */
+			uint32_t	count;		/* T-Read, T-Write, R-Read */
 			char	*data;		/* T-Write, R-Read */
 		};
 		struct
